requirements/requirements-py-3.12.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=atomic_tools --extra=docs --extra=notebook --extra=rest --extra=tests --no-annotate --output-file=requirements/requirements-py-3.12.txt pyproject.toml
#
accessible-pygments==0.0.4
aiida-export-migration-tests==0.9.0
aio-pika==9.4.0
aiormq==6.8.0
alabaster==0.7.13
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.0.0
archive-path==0.4.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
ase==3.22.1
asn1crypto==1.5.1
asttokens==2.4.0
async-generator==1.10
asyncssh~=2.18.0
attrs==23.1.0
babel==2.13.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
circus==0.18.0
click==8.1.7
click-spinner==0.1.10
comm==0.1.4
contourpy==1.1.1
coverage[toml]==7.4.1
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
disk-objectstore==1.2.0
docstring-parser==0.15
docutils==0.20.1
executing==2.0.0
fastjsonschema==2.18.1
flask==2.3.3
flask-cors==3.0.10
flask-restful==0.3.10
fonttools==4.43.1
future==0.18.3
graphviz==0.20.1
greenlet==3.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.3.2
jsonschema[format-nongpl]==3.2.0
jupyter==1.0.0
jupyter-cache==0.6.1
jupyter-client==8.4.0
jupyter-console==6.6.3
jupyter-core==5.4.0
jupyter-events==0.6.3
jupyter-server==2.8.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
kiwipy[rmq]==0.8.4
kiwisolver==1.4.5
latexcodec==2.0.1
mako==1.2.4
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
monty==2023.9.25
mpmath==1.3.0
multidict==6.0.4
myst-nb==1.0.0
myst-parser==2.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2
notebook==6.5.4
notebook-shim==0.2.3
numpy==1.26.1
overrides==7.4.0
packaging==23.2
palettable==3.3.3
pamqp==3.3.0
pandas==2.1.1
pandocfilters==1.5.0
paramiko==3.4.1
parso==0.8.3
pexpect==4.8.0
pg8000==1.30.2
pgsu==0.3.0
pgtest==1.3.2
pickleshare==0.7.5
pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.3.0
plumpy@git+https://github.com/khsrali/plumpy.git@allow-async-upload-download#egg=plumpy
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg[binary]==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pybtex==0.24.0
pycifrw==4.4.5
pycparser==2.21
pydantic==2.4.0
pydantic-core==2.10.0
pydata-sphinx-theme==0.15.1
pygments==2.16.1
pymatgen==2023.10.11
pympler==1.0.1
pymysql==0.9.3
pynacl==1.5.0
pyparsing==3.1.1
pyrsistent==0.19.3
pytest==7.4.2
pytest-asyncio==0.16.0
pytest-benchmark==4.0.0
pytest-cov==4.1.0
pytest-datadir==1.5.0
pytest-regressions==2.5.0
pytest-rerunfailures==12.0
pytest-timeout==2.2.0
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-memcached==1.59
pytray==0.3.4
pytz==2021.3
pyyaml==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
qtpy==2.4.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel-yaml==0.18.2
ruamel-yaml-clib==0.2.8
scipy==1.11.3
scramp==1.4.4
seekpath==1.9.7
send2trash==1.8.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
spglib==2.1.0
sphinx==7.2.6
sphinx-copybutton==0.5.2
sphinx-design==0.5.0
sphinx-intl==2.1.0
sphinx-notfound-page==1.0.0
sphinx-sqlalchemy==0.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.9
sphinxext-rediraffe==0.2.7
sqlalchemy==2.0.23
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing-extensions==4.8.0
tzdata==2023.3
uncertainties==3.1.7
upf-to-json==0.9.5
urllib3==2.0.7
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.4
werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
